fix(competitions): fix routing to phase
